[fix] 입력페이지, 수정페이지 wbs, gantt 적용
