+++ src/pages/wbs/components/wbs-input-table.jsx
@@ -1,250 +1,106 @@
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableRow,
-  TableHead,
-  TextField,
-  Paper,
-  Select,
-  MenuItem,
-  FormControl,
-} from '@mui/material';
-import { memo, useCallback } from 'react';
-import DatePicker from 'react-datepicker';
-import 'react-datepicker/dist/react-datepicker.css';
+import { useState, useEffect } from 'react';
+import { Button, Stack, Container, Grid, Box } from '@mui/material';
+import { useNavigate, useLocation } from 'react-router-dom';
+import { useSelector, useDispatch } from 'react-redux';
+import { setTableData, resetState } from '@/store/wbsSlice';
+import StepperComponent from '@/pages/wbs/components/stepper-component';
+import { PATHS } from '@/routes/paths';
+import { mergeTableDataByRowSpan } from '@/pages/wbs/components/merge-table-data';
+import { postWbs } from '@/services/wbs';
+import WbsFull from '@/pages/wbs/components/wbs-full';
+import GanttFull from '@/pages/wbs/components/gantt-full';
 
-const cellStyle = {
-  border: '1px solid #000',
-  padding: '2px 4px',
+const WbsInputTable = ({ isFullWidth }) => {
+  const tableData = useSelector((state) => state.wbs.tableData);
+  const pjtData = useSelector((state) => state.wbs.pjtData);
+  const memberDatas = useSelector((state) => state.wbs.memberData);
+
+  const [localTableData, setLocalTableData] = useState([]);
+
+  const { pjtSn } = location.state || {};
+  const memberData = memberDatas.map((member, index) => ({
+    userSn: index + 1,
+    userNm: member.name || '',
+    part: member.role || '',
+    role: member.permission || '',
+  }));
+  const memberNames = memberDatas.map((member) => member.name);
+  const ProjectStartDate = new Date(pjtData.startDt);
+  const ProjectEndDate = new Date(pjtData.endDt);
+
+  const wbsData = mergeTableDataByRowSpan(tableData, memberData);
+
+  const finalData = {
+    pjtData,
+    memberData,
+    wbsData,
+  };
+
+  useEffect(() => {
+    if (tableData && tableData.length > 0) {
+      const modifiedData = tableData.map((row) => [...row]);
+      setLocalTableData(modifiedData);
+    }
+  }, [tableData]);
+
+  const handleCellChange = (rowIndex, cellIndex, newValue) => {
+    setLocalTableData((prevData) => {
+      const updatedTable = [...prevData];
+      updatedTable[rowIndex] = [...updatedTable[rowIndex]];
+      updatedTable[rowIndex][cellIndex] = {
+        ...updatedTable[rowIndex][cellIndex],
+        value: newValue,
+      };
+
+      return updatedTable;
+    });
+  };
+
+  const handleSave = () => {
+    dispatch(setTableData(localTableData));
+  };
+
+  const handleBack = () => {
+    handleSave();
+    navigate(PATHS.wbs.basicinfo);
+  };
+
+  return isFullWidth === 100 ?
+      <Box sx={{ display: 'flex', flexDirection: 'row' }}>
+        <Box sx={{ display: 'flex', flexDirection: 'column', flex: 10 }}>
+          <WbsFull
+            tableData={localTableData}
+            handleCellChange={handleCellChange}
+            members={memberNames}
+            editable={false}
+          />
+        </Box>
+      </Box>
+    : <Box sx={{ display: 'flex', flexDirection: 'row' }}>
+        <Box sx={{ display: 'flex', flexDirection: 'column', flex: 6 }}>
+          <WbsFull
+            tableData={localTableData}
+            handleCellChange={handleCellChange}
+            members={memberNames}
+            editable={false}
+          />
+        </Box>
+        <Box
+          sx={{
+            display: 'flex',
+            flexDirection: 'column',
+            overflowX: 'auto',
+            flex: 3,
+          }}
+        >
+          <GanttFull
+            tableData={localTableData}
+            projectStartDate={ProjectStartDate}
+            projectEndDate={ProjectEndDate}
+          />
+          ;
+        </Box>
+      </Box>;
 };
-
-const textFieldStyle = {
-  '& .MuiOutlinedInput-root': {
-    border: 'none',
-    padding: 0,
-    height: '100%',
-    width: '100%',
-    '& fieldset': { border: 'none' },
-  },
-  input: {
-    padding: 0,
-    height: '100%',
-    width: '100%',
-    textAlign: 'center',
-    fontSize: '12px',
-  },
-};
-
-const renderDatePicker = (value, onChange) => (
-  <DatePicker
-    selected={value ? new Date(value) : null}
-    onChange={onChange}
-    customInput={
-      <TextField
-        fullWidth
-        sx={textFieldStyle}
-        inputProps={{ style: { fontSize: '12px' } }}
-      />
-    }
-    dateFormat="yyyy-MM-dd"
-  />
-);
-
-const renderSelectField = (value, onChange, options) => (
-  <FormControl fullWidth variant="outlined">
-    <Select
-      value={value || ''}
-      onChange={(e) => {
-        onChange(e);
-      }}
-      sx={{
-        '& .MuiSelect-select': {
-          padding: '0 !important',
-          fontSize: '12px',
-          textAlign: 'center',
-        },
-        '& .MuiOutlinedInput-notchedOutline': {
-          border: 'none',
-        },
-        '& .MuiSelect-icon': {
-          display: 'none',
-        },
-      }}
-      MenuProps={{
-        PaperProps: {
-          style: {
-            maxHeight: 200,
-            width: 100,
-          },
-        },
-      }}
-    >
-      {options.map((option) => (
-        <MenuItem key={option} value={option}>
-          {option}
-        </MenuItem>
-      ))}
-    </Select>
-  </FormControl>
-);
-
-const renderTextField = (value, onChange) => (
-  <TextField value={value} onChange={onChange} fullWidth sx={textFieldStyle} />
-);
-
-const Cell = memo(
-  ({ cell, rowIndex, cellIndex, handleCellChange, members, editable }) => {
-    const handleChange = useCallback(
-      (e) => handleCellChange(rowIndex, cellIndex, e.target.value),
-      [rowIndex, cellIndex, handleCellChange],
-    );
-    const cellStyleWithEditable = {
-      ...cellStyle,
-      pointerEvents: editable ? 'none' : 'auto', // editable이 false이면 클릭 불가
-    };
-
-    if (cellIndex === 4 || cellIndex === 5) {
-      // 날짜
-      return (
-        <TableCell
-          key={cellIndex}
-          rowSpan={cell.rowSpan}
-          sx={cellStyleWithEditable}
-        >
-          {renderDatePicker(cell.value, (date) =>
-            !editable ? handleCellChange(rowIndex, cellIndex, date) : null,
-          )}
-        </TableCell>
-      );
-    }
-
-    if (cellIndex === 3) {
-      // 엔지니어
-      return (
-        <TableCell
-          key={cellIndex}
-          rowSpan={cell.rowSpan}
-          sx={cellStyleWithEditable}
-        >
-          {renderSelectField(
-            cell.value,
-            !editable ? handleChange : () => {},
-            members,
-          )}
-        </TableCell>
-      );
-    }
-
-    if (cellIndex === 6) {
-      // 상태
-      return (
-        <TableCell
-          key={cellIndex}
-          rowSpan={cell.rowSpan}
-          sx={cellStyleWithEditable}
-        >
-          {renderSelectField(cell.value, !editable ? handleChange : () => {}, [
-            '대기',
-            '진행중',
-            '완료',
-          ])}
-        </TableCell>
-      );
-    }
-
-    // 텍스트 필드
-    return (
-      <TableCell
-        key={cellIndex}
-        rowSpan={cell.rowSpan}
-        sx={cellStyleWithEditable}
-      >
-        {renderTextField(cell.value, !editable ? handleChange : () => {})}
-      </TableCell>
-    );
-  },
-  (prevProps, nextProps) => {
-    return (
-      prevProps.cell.value === nextProps.cell.value &&
-      prevProps.cell.rowSpan === nextProps.cell.rowSpan &&
-      prevProps.members === nextProps.members &&
-      prevProps.editable === nextProps.editable
-    );
-  },
-);
-
-const WbsInputTable = memo(
-  ({ tableData, handleCellChange, members, width, editable }) => {
-    const isFullWidth = width === 100;
-
-    return (
-      <TableContainer
-        component={Paper}
-        sx={{ backgroundColor: '#ffffff', width: `${width}%`, borderRadius: 0 }}
-        elevation={0}
-      >
-        <Table>
-          <colgroup>
-            <col style={{ width: '10%' }} />
-            <col style={{ width: '20%' }} />
-            <col style={{ width: '30%' }} />
-            <col style={{ width: '10%' }} />
-            <col style={{ width: '10%' }} />
-            <col style={{ width: '10%' }} />
-            <col style={{ width: '10%' }} />
-          </colgroup>
-          <TableHead>
-            <TableRow>
-              {[
-                'Part',
-                'Division',
-                'Work',
-                'Engineer',
-                'Start-Date',
-                'Due-Date',
-                'Status',
-              ].map((header) => (
-                <TableCell
-                  key={header}
-                  sx={{
-                    ...cellStyle,
-                    textAlign: 'center',
-                    ...(isFullWidth ?
-                      { fontSize: '14px' }
-                    : { fontSize: '11px', fontWeight: 'fontWeightSemiBold' }),
-                  }}
-                >
-                  {header}
-                </TableCell>
-              ))}
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {tableData.map((row, rowIndex) => (
-              <TableRow key={rowIndex}>
-                {row.map(
-                  (cell, cellIndex) =>
-                    cell && (
-                      <Cell
-                        key={cellIndex}
-                        cell={cell}
-                        rowIndex={rowIndex}
-                        cellIndex={cellIndex}
-                        handleCellChange={handleCellChange}
-                        members={members}
-                        editable={editable}
-                      />
-                    ),
-                )}
-              </TableRow>
-            ))}
-          </TableBody>
-        </Table>
-      </TableContainer>
-    );
-  },
-);
 
 export default WbsInputTable;
